requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
deepface==0.0.79
easydict==1.10
filelock==3.12.0
fire==0.5.0
Flask==2.3.2
flatbuffers==2.0.7
fonttools==4.39.4
gast==0.4.0
gdown==4.7.1
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
imageio==2.29.0
importlib-resources==5.12.0
insightface==0.2.1
itsdangerous==2.1.2
jax==0.4.10
Jinja2==3.1.2
joblib==1.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe==0.9.0.1
ml-dtypes==0.1.0
mtcnn==0.1.1
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
onnx==1.10.1
onnxruntime==1.8.1
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
pandas==2.0.1
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PySocks==1.7.1
pytz==2023.3
PyWavelets==1.4.1
pywin32==305.1
pywin32-ctypes==0.2.0
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
soupsieve==2.4.1
1669632077133/work
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.7.0rc0
tensorflow-gpu==2.7.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.4
wrapt==1.14.1

